requirements/requirements.txt
Django==2.2.9
django-environ==0.8.1
django-user-agents==0.4.0
djangorestframework==3.8.2
mysqlclient==1.3.13
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
ua-parser==0.10.0
user-agents==2.2.0